+++ frontend/src/app/create_customer/rs/page.tsx
@@ -1,0 +1,11 @@
+"use ticket"
+import CreateCustomerRepairshoprScreen from '@/components/screens/create_customer/rs/page'
+import React from 'react'
+
+const CreateCustomerRepairshopr = () => {
+    return (
+        <CreateCustomerRepairshoprScreen />
+    )
+}
+
+export default CreateCustomerRepairshopr