+++ frontend/src/app/create_ticket/rs/page.tsx
@@ -1,0 +1,10 @@
+import CreateTicketRepairshoprScreen from '@/components/screens/departments/create_ticket/rs/page'
+import React from 'react'
+
+const CreateTicketRepairshopr = () => {
+    return (
+        <CreateTicketRepairshoprScreen />
+    )
+}
+
+export default CreateTicketRepairshopr