+++ frontend/src/app/drivers/checklists/[id]/page.tsx
@@ -1,0 +1,11 @@
+import SingleChecklistScreen from '@/components/screens/driver_app/single_checklist/page'
+
+import React from 'react'
+
+const SingleChecklist = () => {
+    return (
+        <SingleChecklistScreen />
+    )
+}
+
+export default SingleChecklist